Add comments and clear the code
