+++ emailProcessingUtils/fetchThread.js
@@ -1,52 +1,56 @@
-const Imap = require('imap');
+// Libraries
 const simpleParser = require('mailparser').simpleParser;
-require('dotenv').config();
+const Imap = require('imap');
+
+// Config
+const imapConfig = require('../imapConfig');
 
 // IMAP configuration
-const imap = new Imap({
-  user: process.env.EMAIL_USER,
-  password: process.env.EMAIL_PASS,
-  host: 'imap.gmail.com',
-  port: 993,
-  tls: true,
-  tlsOptions: {
-    rejectUnauthorized: false,
-  },
-});
+const imap = new Imap(imapConfig);
 
+// Function to open the inbox
 function openInbox(cb) {
   imap.openBox('INBOX', true, cb);
 }
 
+// Function to fetch the email thread using the threadId
 async function fetchThread(threadId) {
   return new Promise((resolve, reject) => {
     let threadArray = [];
 
+    // When the IMAP connection is ready, open the inbox
     imap.once('ready', function () {
       openInbox(async function (err, box) {
         if (err) reject(err);
 
+        // Search for messages in the thread using the threadId
         imap.search([['X-GM-THRID', `${threadId}`]], async function (err, results) {
           if (err) reject(err);
 
+          // Fetch the messages in the thread
           const f = imap.fetch(results, { bodies: '' });
           let messageCount = results.length;
           let processedMessages = 0;
 
+          // Process each message in the thread
           f.on('message', function (msg, seqno) {
             msg.on('body', async function (stream, info) {
               try {
+                // Parse the message using the simpleParser library
                 const parsed = await simpleParser(stream);
 
+                // Create a message object with the sender's email address and content
                 const msg = {
                   from: parsed.from.value[0].address,
                   content: parsed.text
                 }
+                // Add the message object to the thread array
                 threadArray.push(msg);
               } catch (err) {
                 reject(err);
               } finally {
                 processedMessages++;
+                // Once all messages are processed, close the IMAP connection and resolve the thread array
                 if (processedMessages === messageCount) {
                   imap.end();
                   resolve(threadArray);
@@ -55,6 +59,7 @@ async function fetchThread(threadId) {
             });
           });
 
+          // Handle fetch errors
           f.once('error', function (err) {
             reject('Fetch error: ' + err);
           });
@@ -62,16 +67,15 @@ async function fetchThread(threadId) {
       });
     });
 
+    // Handle IMAP connection errors
     imap.once('error', function (err) {
       reject(err);
     });
 
-    imap.once('end', function () {
-      console.log('Connection ended (fetchThread)');
-    });
-
+    // Connect to the IMAP server
     imap.connect();
   });
 }
 
+// Export the fetchThread function
 module.exports = fetchThread;
