+++ imapConfig.js
@@ -1,0 +1,15 @@
+require('dotenv').config();
+
+const config = {
+    user: process.env.EMAIL_USER,
+    password: process.env.EMAIL_PASS,
+    host: 'imap.gmail.com',
+    port: 993,
+    tls: true,
+    tlsOptions: {
+      rejectUnauthorized: false,
+    },
+  };
+  
+  module.exports = config;
+
